DISCO-1896 Respond to each request with an authentication indicator
