+++ src/main/java/com/dnastack/ddapfrontend/route/BeaconResource.java
@@ -9,7 +9,7 @@ import com.dnastack.ddapfrontend.client.dam.DamResource;
 import com.dnastack.ddapfrontend.client.dam.DamResourceList;
 import com.dnastack.ddapfrontend.client.dam.DamView;
 import com.dnastack.ddapfrontend.model.BeaconRequestModel;
-import com.dnastack.ddapfrontend.proxy.SetBearerTokenFromCookieGatewayFilterFactory;
+import com.dnastack.ddapfrontend.security.UserTokenStatusFilter;
 import lombok.Value;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.server.reactive.ServerHttpRequest;
@@ -20,7 +20,10 @@ import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 import static java.lang.String.format;
@@ -66,7 +69,7 @@ class BeaconResource {
             ServerHttpRequest request) {
 
         // find beacons under resourceId in DAM config
-        Optional<String> damToken = SetBearerTokenFromCookieGatewayFilterFactory.extractDamToken(request);
+        Optional<String> damToken = UserTokenStatusFilter.extractDamToken(request);
         if (!damToken.isPresent()) {
             return Flux.error(new IllegalArgumentException("Authorization is required")); // TODO make this return a 401
         }
